requirements.txt
Flask==2.3.2
gunicorn==23.0.0
gspread==6.1.4
google-auth==2.40.3
selenium==4.15.2
requests==2.32.5
python-dotenv==1.0.0
